requirements.txt
arabic-reshaper==3.0.0
asgiref==3.2.10
asn1crypto==1.5.1
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.4
colorama==0.4.6
cryptography==41.0.1
cssselect2==0.7.0
datetime-truncate==1.1.1
Django==3.1
django-admin-charts==1.3.0
django-admin-thumbnails==0.2.8
django-bower==5.2.0
django-chartjs==2.3.0
django-jazzmin==2.6.0
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-session-timeout==0.1.0
fonttools==4.40.0
html5lib==1.1
idna==3.4
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
oscrypto==1.3.0
Pillow==9.5.0
psycopg2==2.9.6
pycparser==2.21
pydyf==0.7.0
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
pypdf==3.12.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
Unidecode==1.3.6
uritools==4.0.1
urllib3==2.0.3
weasyprint==59.0
webencodings==0.5.1
xhtml2pdf==0.2.11
zopfli==0.2.2
